+++ src/main/java/com/example/hits/presentation/dto/taskanswer/TaskAnswerModel.java
@@ -17,9 +17,9 @@ public class TaskAnswerModel {
 
     private UUID id = UUID.randomUUID();
 
-    private Integer score = null;
+    private Float score = null;
 
-    private Integer maxScore = 100;
+    private Float maxScore = 100f;
 
     private LocalDateTime submittedAt = null;
